backend/requirements.txt
# Web framework
fastapi==0.115.0
uvicorn[standard]==0.30.6
websockets==13.0

# Database
sqlalchemy==2.0.35
asyncpg==0.29.0
aiosqlite==0.20.0
alembic==1.13.3

# Cache
redis==5.1.0

# Config
pydantic==2.9.2
pydantic-settings==2.5.2

# Testing
pytest==8.3.3

# System monitoring
psutil==6.0.0
paramiko==3.4.0  # For SSH remote monitoring
prometheus-client==0.20.0  # For Prometheus integration
httpx==0.27.2

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Code quality
black==24.10.0
flake8==7.1.1

# Multipart support for OAuth2 form
python-multipart==0.0.7
